fix(web): remove prefetch on paginated tables
RTK has an issue with subscriptions on prefetch queries, which causes them to be re-fetched even when subscriber count is 0. This caused errors when an entity is removed and its related data is prefetched and cached (for example, instrument and instrument prices table) - RTK would attempt to re-fetch data which no component is subscribed to.